Add Operation Handler
A new class, Operation Handler, is introduced to simplify services
focused on executing operations in a transaction.

This is mostly useful when operation visitors can be pre-configured,
allowing the calling service in turn to not be aware of the visitors.
While using a bare Transaction, the visitors are passed on at runtime,
while the operation handler allows to have them attached before the
operations are handled.